requirements.txt
django
psycopg2-binary
pillow
django-ninja
django-silk
django-ninja-simple-jwt